feat: centralize transaction metrics helpers
