+++ controllers/reportController.js
@@ -1,24 +1,6 @@
 const supabase = require('../supabaseClient');
 const { assertSupabase } = require('../supabaseClient');
-
-const DAY = 24 * 60 * 60 * 1000;
-function parseISOorNull(s) {
-  try {
-    if (!s) return null;
-    const d = new Date(s);
-    return isNaN(d) ? null : d;
-  } catch (_) {
-    return null;
-  }
-}
-function periodFromQuery(q) {
-  const to = parseISOorNull(q.to) || new Date();
-  const from = parseISOorNull(q.from) || new Date(to.getTime() - 30 * DAY);
-  return { from, to };
-}
-function asISOdate(d) {
-  return d.toISOString();
-}
+const { periodFromQuery, iso, aggregate } = require('../services/transacoesMetrics');
 
 exports.resumo = async (req, res) => {
   if (!assertSupabase(res)) return;
@@ -26,81 +8,20 @@ exports.resumo = async (req, res) => {
   const { data, error } = await supabase
     .from('transacoes')
     .select('cpf,cliente_nome,plano,valor_bruto,valor_final')
-    .gte('created_at', from.toISOString())
-    .lte('created_at', to.toISOString());
+    .gte('created_at', iso(from))
+    .lte('created_at', iso(to));
   if (error) return res.status(500).json({ error: error.message });
 
-  const totalTransacoes = data.length;
-  let totalBruto = 0;
-  let totalLiquido = 0;
-  let totalDescontos = 0;
-  const planos = {};
-  const clientes = {};
-
-  for (const tx of data) {
-    const bruto = Number(tx.valor_bruto) || 0;
-    const liquido = Number(tx.valor_final) || 0;
-    const desconto = bruto - liquido;
-
-    totalBruto += bruto;
-    totalLiquido += liquido;
-    totalDescontos += desconto;
-
-    if (!planos[tx.plano])
-      planos[tx.plano] = {
-        plano: tx.plano,
-        qtd: 0,
-        bruto: 0,
-        descontos: 0,
-        liquido: 0,
-      };
-    const p = planos[tx.plano];
-    p.qtd++;
-    p.bruto += bruto;
-    p.descontos += desconto;
-    p.liquido += liquido;
-
-    if (!clientes[tx.cpf])
-      clientes[tx.cpf] = {
-        cpf: tx.cpf,
-        nome: tx.cliente_nome,
-        qtd: 0,
-        bruto: 0,
-        descontos: 0,
-        liquido: 0,
-      };
-    const c = clientes[tx.cpf];
-    c.qtd++;
-    c.bruto += bruto;
-    c.descontos += desconto;
-    c.liquido += liquido;
-  }
-
-  const round = (n) => Number(n.toFixed(2));
-  const porPlano = Object.values(planos).map((p) => ({
-    plano: p.plano,
-    qtd: p.qtd,
-    bruto: round(p.bruto),
-    descontos: round(p.descontos),
-    liquido: round(p.liquido),
-  }));
-  const porCliente = Object.values(clientes).map((c) => ({
-    cpf: c.cpf,
-    nome: c.nome,
-    qtd: c.qtd,
-    bruto: round(c.bruto),
-    descontos: round(c.descontos),
-    liquido: round(c.liquido),
-  }));
+  const metrics = aggregate(data, { from, to });
 
   return res.json({
-    periodo: { from: asISOdate(from), to: asISOdate(to) },
-    totalTransacoes,
-    totalBruto: round(totalBruto),
-    totalDescontos: round(totalDescontos),
-    totalLiquido: round(totalLiquido),
-    porPlano,
-    porCliente,
+    periodo: { from: iso(from), to: iso(to) },
+    totalTransacoes: metrics.qtdTransacoes,
+    totalBruto: metrics.bruto,
+    totalDescontos: metrics.descontos,
+    totalLiquido: metrics.liquido,
+    porPlano: metrics.porPlano,
+    porCliente: metrics.porCliente,
   });
 };
 
@@ -112,8 +33,8 @@ exports.csv = async (req, res) => {
     .select(
       'id,created_at,cpf,cliente_nome,plano,valor_bruto,desconto_aplicado,valor_final,origem'
     )
-    .gte('created_at', from.toISOString())
-    .lte('created_at', to.toISOString());
+    .gte('created_at', iso(from))
+    .lte('created_at', iso(to));
 
   if (req.query.cpf) q.eq('cpf', req.query.cpf);
   if (req.query.plano) q.eq('plano', req.query.plano);
@@ -152,10 +73,6 @@ exports.csv = async (req, res) => {
 };
 
 module.exports = {
-  DAY,
-  parseISOorNull,
-  periodFromQuery,
-  asISOdate,
   resumo: exports.resumo,
   csv: exports.csv,
 };
